@@ -1,6 +1,6 @@
 {
   "name": "ngx-malihu-scrollbar",
-  "version": "9.0.0",
+  "version": "10.0.0-beta.0",
   "description": "Angular scrollbar customization using Malihu jQuery Custom Scrollbar plugin",
   "homepage": "https://github.com/jfcere/ngx-malihu-scrollbar",
   "license": "MIT",
@@ -21,10 +21,11 @@
   "dependencies": {
     "@types/mcustomscrollbar": "^2.8.32",
     "jquery": "^3.4.1",
-    "malihu-custom-scrollbar-plugin": "^3.1.5"
+    "malihu-custom-scrollbar-plugin": "^3.1.5",
+    "tslib": "^2.0.0"
   },
   "peerDependencies": {
-    "@angular/core": "^8.0.0 || ^9.0.0",
-    "zone.js": "^0.9.0 || ^0.10.0"
+    "@angular/core": "^10.0.0",
+    "zone.js": "^0.10.3"
   }
 }
