@@ -1,6 +1,6 @@
 {
   "name": "ngx-malihu-scrollbar",
-  "version": "9.0.0",
+  "version": "10.0.0-beta.0",
   "description": "Angular scrollbar customization using Malihu jQuery Custom Scrollbar plugin",
   "homepage": "https://github.com/jfcere/ngx-malihu-scrollbar",
   "license": "MIT",
@@ -36,47 +36,45 @@
     "publish:lib": "npm publish ./dist/lib"
   },
   "dependencies": {
-    "@angular/animations": "^9.0.0",
-    "@angular/common": "^9.0.0",
-    "@angular/compiler": "^9.0.0",
-    "@angular/core": "^9.0.0",
-    "@angular/forms": "^9.0.0",
-    "@angular/platform-browser": "^9.0.0",
-    "@angular/platform-browser-dynamic": "^9.0.0",
-    "@angular/router": "^9.0.0",
+    "@angular/animations": "^10.1.1",
+    "@angular/common": "^10.1.1",
+    "@angular/compiler": "^10.1.1",
+    "@angular/core": "^10.1.1",
+    "@angular/forms": "^10.1.1",
+    "@angular/platform-browser": "^10.1.1",
+    "@angular/platform-browser-dynamic": "^10.1.1",
+    "@angular/router": "^10.1.1",
     "ngx-malihu-scrollbar": "file:./lib",
-    "rxjs": "^6.5.4",
-    "tslib": "^1.10.0",
+    "rxjs": "~6.6.0",
+    "tslib": "^2.0.0",
     "zone.js": "^0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.900.0",
-    "@angular-devkit/build-ng-packagr": "~0.900.0",
-    "@angular/cli": "^9.0.0",
-    "@angular/compiler-cli": "^9.0.0",
-    "@angular/language-service": "^9.0.0",
+    "@angular-devkit/build-angular": "~0.1001.1",
+    "@angular-devkit/build-ng-packagr": "~0.1001.1",
+    "@angular/cli": "^10.1.1",
+    "@angular/compiler-cli": "^10.1.1",
+    "@angular/language-service": "^10.1.1",
     "@types/jasmine": "~3.5.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^12.11.1",
     "angular-cli-ghpages": "^0.5.3",
-    "codelyzer": "^5.1.2",
+    "codelyzer": "^6.0.0",
     "coveralls": "^2.13.3",
     "cpx": "^1.5.0",
-    "jasmine-core": "~3.5.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.3.0",
+    "jasmine-core": "~3.6.0",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "~5.0.0",
     "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "~2.1.0",
-    "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.2",
+    "karma-coverage-istanbul-reporter": "~3.0.2",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
     "karma-junit-reporter": "^1.2.0",
     "linklocal": "^2.8.2",
-    "ng-packagr": "^9.0.0",
-    "rimraf": "^2.6.3",
+    "ng-packagr": "^10.1.0",
+    "rimraf": "^2.7.0",
     "ts-node": "~8.3.0",
-    "tslint": "~5.18.0",
-    "tslint-eslint-rules": "^5.4.0",
-    "tslint-microsoft-contrib": "^6.2.0",
-    "typescript": "~3.7.5"
+    "tslint": "~6.1.0",
+    "typescript": "~4.0.2"
   }
 }
